Transfert du code de MainActivity vers LoginActivity
