+++ app/src/main/res/values/strings.xml
@@ -1,13 +1,12 @@
 <resources>
     <string name="app_name">toulouse-0218-java-rateeverything</string>
     <string name="pseudo">Pseudo</string>
-    <string name="login">Login</string>
     <string name="create">Create</string>
     <string name="modifier_l_avatar">Modifier l\'avatar</string>
     <string name="return1">Return</string>
     <string name="password">Password</string>
-    <string name="password">Password</string>
     <string name="create_an_account">Create An Account</string>
-    <string name="pseudo">Pseudo</string>
     <string name="sign_in">Sign In</string>
+    <string name="login">Login</string>
+    <string name="page_login">page Login</string>
 </resources>
